Renamed accountActions to authActions
