+++ app/containers/LoginPrivateKey/index.js
@@ -4,7 +4,7 @@ import { compose } from 'recompose'
 import LoginPrivateKey from './LoginPrivateKey'
 import withActions from '../../hocs/api/withActions'
 import withFailureNotification from '../../hocs/withFailureNotification'
-import { wifLoginActions } from '../../actions/accountActions'
+import { wifLoginActions } from '../../actions/authActions'
 
 const mapActionsToProps = (actions) => ({
   loginWithPrivateKey: (wif) => actions.request({ wif })
