+++ app/containers/TransactionHistory/index.js
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux'
 import { compose } from 'recompose'
 
 import withNetworkData from '../../hocs/withNetworkData'
-import withAccountData from '../../hocs/withAccountData'
+import withAuthData from '../../hocs/withAuthData'
 import { syncTransactionHistory, getIsLoadingTransactions } from '../../modules/transactions'
 import { getTransactions } from '../../modules/wallet'
 import TransactionHistory from './TransactionHistory'
@@ -23,5 +23,5 @@ const mapDispatchToProps = dispatch => bindActionCreators(actionCreators, dispat
 export default compose(
   connect(mapStateToProps, mapDispatchToProps),
   withNetworkData(),
-  withAccountData()
+  withAuthData()
 )(TransactionHistory)
