+++ app/hocs/withAuthData.js
@@ -2,7 +2,7 @@
 import { isEmpty, zipObject, mapValues, invert } from 'lodash'
 
 import withData from './api/withData'
-import accountActions from '../actions/accountActions'
+import authActions from '../actions/authActions'
 
 type Mapping = {
   address?: string,
@@ -14,10 +14,10 @@ type Mapping = {
 const keys: Array<string> = ['address', 'publicKey', 'wif', 'signingFunction']
 const defaultMapping: Mapping = zipObject(keys, keys)
 
-export default function withAccountData (mapping: Mapping = defaultMapping) {
-  return withData(accountActions, (account) => {
-    if (!isEmpty(account)) {
-      return mapValues(invert(mapping), (field) => account[field])
+export default function withAuthData (mapping: Mapping = defaultMapping) {
+  return withData(authActions, (auth) => {
+    if (!isEmpty(auth)) {
+      return mapValues(invert(mapping), (field) => auth[field])
     } else {
       return {}
     }
